requirements.txt
fastmcp
sentence-transformers
faiss-cpu
openai
anthropic
google-generativeai
playwright
python-dotenv
configparser
Pillow
torch
torchvision
beautifulsoup4
pyyaml
scipy
numpy
